Support cross compilation of arm64 on non-arm64 systems
Built slowdebug/fastdebug/release for arm64
Built release for x86_64 for regressions

contains:
- make file changes to support cross-compilation
- replaced use of deprecated API (JRSCopyOSVersion) / UNTESTED

here's some test code if you want to validate the change without hardware:

````

int encode (int major, int minor)
{
    major *= 10;
    int tmp = minor;
    while (tmp > 9) {
        tmp /= 10;
        major *= 10;
    }
    return major + minor;
}

int old_encode(char* ver)
{
    int len = strlen(ver);
    int v = 0;

    // Third char must be a '.'
    if (len >= 3 && ver[2] == '.') {
        int i;

        v = (ver[0] - '0') * 10 + (ver[1] - '0');
        for (i = 3; i < len && isdigit(ver[i]); ++i) {
            v = v * 10 + (ver[i] - '0');
        }
    }

    return v;
}

int main (int argc, char* argv [])
{
    int major, minor;
    sscanf(argv[1], "%d.%d", &major, &minor);
    int new_result = encode (major, minor);
    int old_result = old_encode (argv[1]);
    printf("new encode = %s becomes %d\n", argv[1], new_result);
    printf("old encode = %s becomes %d\n", argv[1], old_result);
    return 0;
}

Related work items: #5406, #6627

+++ src/java.desktop/macosx/native/libawt_lwawt/awt/OSVersion.h
@@ -25,5 +25,5 @@
 
 // Support for detecting Mac OS X versions
 
-double getOSXMajorVersion();
+int getOSXMajorVersion();
 BOOL isSnowLeopardOrLower();
